factor out Utils for higher-order recursion schemes
