+++ src/Extract.hs
@@ -1,36 +1,16 @@
-{-# LANGUAGE GADTs              #-}
-{-# LANGUAGE KindSignatures     #-}
-{-# LANGUAGE DeriveFunctor      #-}
-{-# LANGUAGE LambdaCase         #-}
-{-# LANGUAGE TypeOperators      #-}
-{-# LANGUAGE RankNTypes         #-}
-{-# LANGUAGE TypeFamilies       #-}
-{-# LANGUAGE FlexibleContexts   #-}
+{-# LANGUAGE GADTs        #-}
+{-# LANGUAGE TypeFamilies #-}
+{-# LANGUAGE LambdaCase   #-}
 
 module Extract where
 
-import Control.Natural
 import Data.Functor.Const
 import qualified Data.List.NonEmpty as NonEmpty
 
 import RefinedAst
 import Syntax
+import Utils
 
--- Utils that should be moved eventually --
-
-class HFunctor (h :: (* -> *) -> * -> *) where
-  hfmap :: (f ~> g) -> h f ~> h g
-
-class HFoldable (h :: (* -> *) -> * -> *) where
-  hfoldMap :: Monoid m => (forall b. f b -> m) -> h f a -> m
-
-class HFunctor (HBase r) => HRecursive (r :: * -> *) where
-  type HBase r :: (* -> *) -> * -> *
-
-  hproject :: r ~> HBase r r
-
-  hcata :: (HBase r f ~> f) -> r ~> f
-  hcata alg = alg . hfmap (hcata alg) . hproject
 
 -- Non-recursive Expression type and necessary instances
